backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
openai>=1.10.0
python-dotenv>=1.0.0
httpx>=0.26.0
python-multipart>=0.0.6
aiofiles>=23.2.1
PyJWT>=2.8.0
psycopg2-binary>=2.9.0
